[DDW-321] Separates API errors into sub-directories by category
