+++ source/renderer/app/api/wallets/errors.js
@@ -7,6 +7,16 @@ const messages = defineMessages({
     defaultMessage: '!!!Wallet you are trying to restore already exists.',
     description: '"Wallet you are trying to restore already exists." error message.'
   },
+  walletAlreadyImportedError: {
+    id: 'api.errors.WalletAlreadyImportedError',
+    defaultMessage: '!!!Wallet you are trying to import already exists.',
+    description: '"Wallet you are trying to import already exists." error message.'
+  },
+  walletFileImportError: {
+    id: 'api.errors.WalletFileImportError',
+    defaultMessage: '!!!Wallet could not be imported, please make sure you are providing a correct file.',
+    description: '"Wallet could not be imported, please make sure you are providing a correct file." error message.'
+  },
 });
 
 export class WalletAlreadyRestoredError extends LocalizableError {
@@ -14,6 +24,24 @@ export class WalletAlreadyRestoredError extends LocalizableError {
     super({
       id: messages.walletAlreadyRestoredError.id,
       defaultMessage: messages.walletAlreadyRestoredError.defaultMessage,
+    });
+  }
+}
+
+export class WalletAlreadyImportedError extends LocalizableError {
+  constructor() {
+    super({
+      id: messages.walletAlreadyImportedError.id,
+      defaultMessage: messages.walletAlreadyImportedError.defaultMessage,
+    });
+  }
+}
+
+export class WalletFileImportError extends LocalizableError {
+  constructor() {
+    super({
+      id: messages.walletFileImportError.id,
+      defaultMessage: messages.walletFileImportError.defaultMessage,
     });
   }
 }
